+++ src/app/db/models/download.py
@@ -16,8 +16,8 @@ class DOWNLOADER(Base):
 
     ID: Mapped[int] = mapped_column(Integer, Sequence("ID"), primary_key=True)
     NAME: Mapped[str] = mapped_column(String(255))
-    ENABLED: Mapped[int] = mapped_column(Integer)
-    TYPE: Mapped[str] = mapped_column(String(255))
+    ENABLED: Mapped[int] = mapped_column(Integer, index=True)
+    TYPE: Mapped[str] = mapped_column(String(255), index=True)
     TRANSFER: Mapped[int] = mapped_column(Integer)
     ONLY_NEXUS_MEDIA: Mapped[int] = mapped_column(Integer)
     MATCH_PATH: Mapped[int] = mapped_column(Integer)